feat: lib.rs
