+++ Cargo.toml
@@ -3,6 +3,11 @@ name = "sabi"
 version = "0.1.0"
 edition = "2024"
 
+[lib]
+name = "sabi"
+path = "src/lib.rs"
+crate-type = ["rlib"]
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
