--- assets/acts/Chapter 1/1.sabi
@@ -1,64 +1,0 @@
-SCENE intro
-    (GUI textbox changes to "TEXTBOX_NASTYA")
-    (GUI namebox changes to "NAMEBOX")
-    (Background changes to "main_classroom_noon")
-    (Nayu appears concerned)
-    MC: "Welcome to the new scene system!"
-    Nayu: (neutral) "This is the intro scene."
-
-    (Scene "main" begins)
-CURTAIN
-
-SCENE main
-    (Background changes to "main_classroom_day")
-    MC: "This is the main scene!"
-    Nayu: (happy) "You can now switch between scenes easily."
-
-    (Nayu disappears)
-    (Scene "characterspawning" begins)
-CURTAIN
-
-SCENE characterspawning
-    MC: "We are making some tests on character handling!"
-    (Nayu appears)
-    Nayu: "I just appeared with my default emotion, which is neutral :|"
-    Nayu: (happy) "But this seems to be working, so it is good"
-    (Nayu disappears)
-    Nayu: "I disappeared"
-    Nayu: (sad) "From now on, until I reappear, emotion changes will not be displayed nor applied :("
-    (Nayu fade in)
-    Nayu: "I just learned how to fade in into the scene and: guess what?"
-    (Nayu fade out)
-    
-    (Scene "characterdirection" begins)
-CURTAIN
-
-SCENE characterdirection
-    (Nayu appears)
-    Nayu: "Default direction of sprite images must be towards right."
-    (Nayu looks left)
-    Nayu: "In this way I can coherently change my direction during the scene"
-    (Nayu looks right)
-    Nayu: "Use 'looks'/'looking' directive to change my direction"
-    (Nayu fade out)
-    (Scene "charactersmovement" begins)
-CURTAIN
-
-SCENE charactersmovement
-    (Nayu appears invisible left happy)
-    Nayu: "It seems like I'm not here..."
-    (Nayu moves center)
-    Nayu: "But here I am! Ready to move left and right on the screen"
-    (Nayu moves left)
-    Nayu: "Speaking of which... I think I'm gonna leave again"
-    (Nayu moves invisible right)
-    (Nayu fade out)
-    (Scene "epilogue" begins)
-CURTAIN
-
-SCENE epilogue
-    MC: "Normally, this scene won't be reached directly."
-    Nayu: "But it was accessed via scene commands!"
-
-    (Act "2" begins)
-CURTAIN
