--- assets/acts/Chapter 1/2.sabi
@@ -1,13 +1,0 @@
-SCENE beginning
-    (Background changes to "main_classroom_night")
-    MC: "Welcome to Act 2!"
-    Nayu: (concerned) "This is a different act entirely."
-    MC: "The system automatically started with the first scene in this act."
-
-    (Scene "end" begins)
-CURTAIN
-
-SCENE end
-    MC: "This is the last scene in Act 2."
-    Nayu: (neutral) "You could jump here with a scene change command."
-CURTAIN
