+++ assets/acts/examples/basic_startup.sabi
@@ -1,0 +1,10 @@
+SCENE basic_startup
+    (GUI textbox changes to "TEXTBOX_NASTYA")
+    (GUI namebox changes to "NAMEBOX")
+    (Background changes to "main_classroom_noon")
+    (Nayu appears concerned)
+    MC: "Welcome to Sabi! A Visual Novel engine written in Rust and Bevy."
+    Nayu: (neutral) "This is the intro scene."
+    Nayu: "You're viewing the 'basic_startup' example."
+    Nayu: "Check the other examples to dive in into Sabi library."
+CURTAIN
